Visualise multiple algorithm results
The visualisation is revised be able to visualise multiple algorithm
results.

The code is also refactored not to include hidden checkboxes
that handle showing patterns on and off, but the buttons handle
the logic on their own. Also some other refactoring where the
data is processed first to be in a convenient form to then
draw everything with D3.