Fix bug when not selecting the first track of a file
There was a bug that tempo messages were lost if the user did not
choose to have the first track of a file in the point set,
resulting in all notes happening at time 0. Filtering not chosen
tracks is changed to avoid this: now the velocity of all notes
in not chosen tracks is turned to 0.